correcao no valor das tabelas

+++ src/app/models/TravelExpenses.js
@@ -10,7 +10,7 @@ class TravelExpenses extends Model {
         name_establishment: Sequelize.STRING,
         expense_description: Sequelize.STRING,
         dfe: Sequelize.STRING,
-        value: Sequelize.DECIMAL,
+        value: Sequelize.INTEGER,
         proof_img: Sequelize.STRING,
       },
       {
